// src/types/RootStackParamList.tsx
import {UserType} from './UserType';
import {ReviewType} from './ReviewType';
import {DoctorType} from './DoctorType';
import {DiagnosticType} from './DiagnosticType';

export type RootStackParamList = {
  FAQ: undefined;
  Filter: undefined;
  SignIn: undefined;
  SignUp: undefined;
  Search: undefined;
  LogOut: undefined;
  Languages: undefined;
  MyProfile: undefined;
  Gravatar: undefined;
  Web: {url: string; title: string};
  UserCreationProcess: {user: UserType};
  DoctorDetails: {doctor: DoctorType};
  DashboardVersion: undefined;
  MyAddress: undefined;
  Onboarding: undefined;
  EditProfile: undefined;
  OrderFailed: undefined;
  AddANewCard: undefined;
  OrderHistory: undefined;
  LeaveAReview: undefined;
  NewPassword: undefined;
  MyPromocodes: undefined;
  TabNavigator: undefined;
  Notifications: undefined;
  LeaveAReviews: undefined;
  PaymentMethod: undefined;
  ForgotPassword: undefined;
  TrackYourOrder: {orderId: string};
  AddANewAddress: undefined;
  OrderSuccessful: undefined;
  MakeAnAppointment: {doctor: DoctorType} | undefined;
  ConfirmationCode: undefined;
  MyTestsAndDiagnostics: undefined;
  VerifyYourEmail: undefined;
  PrivacyPolicy: undefined;
  LanguageList: undefined;
  Reviews: {reviews: ReviewType[]};
  MyAppointments: undefined;
  AppointmentSuccess: undefined;
  MyPromocodesEmpty: undefined;
  SignUpAccountCreated: undefined;
  CheckoutPaymentMethod: undefined;
  VerifyYourPhoneNumber: undefined;
  CheckoutShippingDetails: undefined;
  DoctorListVersion: undefined;
  Filters: undefined;
  ConfirmationCodeEmail: undefined;
  ForgotPasswordSentEmail: undefined;
  MyDoctors: undefined;
  Biography: {doctor: DoctorType};
  ChangeCurrency: undefined;
  DoctorsList: undefined;
  VerifyPhoneNumber: {phoneNumber: string};
  NumberChanged: undefined;
  AccountCreated: undefined;
  PasswordChanged: undefined;
  DirectMessages: {history: any};
  AccountCreatedFailed: undefined;
  DiagnosticDetails: {item: DiagnosticType};
};